+++ tests/testthat/test-set_host.R
@@ -2,7 +2,7 @@ test_gitstats <- create_gitstats()
 
 test_that("Set connection returns appropriate messages", {
   expect_snapshot(
-    set_connection(
+    set_host(
       gitstats_obj = test_gitstats,
       api_url = "https://api.github.com",
       token = Sys.getenv("GITHUB_PAT"),
@@ -10,7 +10,7 @@ test_that("Set connection returns appropriate messages", {
     )
   )
   expect_snapshot(
-    test_gitstats %>% set_connection(
+    test_gitstats %>% set_host(
       api_url = "https://gitlab.com/api/v4",
       token = Sys.getenv("GITLAB_PAT_PUBLIC"),
       orgs = c("mbtests")
@@ -21,7 +21,7 @@ test_that("Set connection returns appropriate messages", {
 test_that("When empty token for GitHub, GitStats pulls default token", {
   expect_snapshot(
     test_gitstats <- create_gitstats() %>%
-      set_connection(
+      set_host(
         api_url = "https://api.github.com",
         orgs = c("openpharma", "r-world-devs")
       )
@@ -32,7 +32,7 @@ test_that("When empty token for GitLab, GitStats pulls default token", {
   expect_snapshot(
     withr::with_envvar(new = c("GITLAB_PAT" = Sys.getenv("GITLAB_PAT_PUBLIC")), {
       test_gitstats <- create_gitstats() %>%
-        set_connection(
+        set_host(
           api_url = "https://gitlab.com/api/v4",
           orgs = "mbtests"
         )
@@ -44,7 +44,7 @@ test_that("Error shows if organizations are not specified and host is not passed
   test_gitstats <- create_gitstats()
   expect_snapshot_error(
     test_gitstats %>%
-      set_connection(
+      set_host(
         api_url = "https://api.github.com",
         token = Sys.getenv("GITHUB_PAT")
       )
@@ -59,7 +59,7 @@ test_that("Error shows, when wrong input is passed when setting connection and h
   test_gitstats <- create_gitstats()
 
   expect_snapshot_error(
-    set_connection(
+    set_host(
       gitstats_obj = test_gitstats,
       api_url = "https://avengers.com",
       token = Sys.getenv("GITLAB_PAT_PUBLIC")
@@ -73,12 +73,12 @@ test_that("Error pops out, when two clients of the same url api are passed as in
   expect_snapshot(
     error = TRUE,
     test_gitstats %>%
-      set_connection(
+      set_host(
         api_url = "https://api.github.com",
         token = Sys.getenv("GITHUB_PAT"),
         orgs = "pharmaverse"
       ) %>%
-      set_connection(
+      set_host(
         api_url = "https://api.github.com",
         token = Sys.getenv("GITHUB_PAT"),
         orgs = "openpharma"
@@ -89,7 +89,7 @@ test_that("Error pops out, when two clients of the same url api are passed as in
 test_that("`Org` name is not passed to the object if it does not exist", {
   expect_snapshot(
     test_gitstats <- create_gitstats() %>%
-      set_connection(
+      set_host(
         api_url = "https://api.github.com",
         token = Sys.getenv("GITHUB_PAT"),
         orgs = c("openparma")
@@ -98,7 +98,7 @@ test_that("`Org` name is not passed to the object if it does not exist", {
 
   expect_snapshot(
     test_gitstats <- create_gitstats() %>%
-      set_connection(
+      set_host(
         api_url = "https://gitlab.com/api/v4",
         token = Sys.getenv("GITLAB_PAT_PUBLIC"),
         orgs = c("openparma", "mbtests")
@@ -107,7 +107,7 @@ test_that("`Org` name is not passed to the object if it does not exist", {
 
   expect_snapshot(
     test_gitstats <- create_gitstats() %>%
-      set_connection(
+      set_host(
         api_url = "https://api.github.com",
         token = Sys.getenv("GITHUB_PAT"),
         orgs = c("openpharma", "r_world_devs")